# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWC_test doc API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "description": "Endpoint to create a new booking for a resource by specifying the resource ID, user ID, start time, and end time.",
                "operationId": "0a2e7491d776c6a5af9518087ab96f0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "resource_id": {
                                        "description": "The ID of the resource.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "description": "The ID of the user.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "description": "Start time of the booking.",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01 12:00:00"
                                    },
                                    "end_time": {
                                        "description": "End time of the booking.",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01 14:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "The ID of the created booking.",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "resource_id": {
                                            "description": "The ID of the resource.",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "description": "The ID of the user.",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_time": {
                                            "description": "Start time of the booking.",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01 12:00:00"
                                        },
                                        "end_time": {
                                            "description": "End time of the booking.",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01 14:00:00"
                                        },
                                        "created_at": {
                                            "description": "Timestamp when the booking was created.",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01 12:00:00"
                                        },
                                        "updated_at": {
                                            "description": "Timestamp when the booking was last updated.",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-01 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "resource_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The resource id field is required."
                                                    }
                                                },
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The user id field is required."
                                                    }
                                                },
                                                "start_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The start time field is required."
                                                    }
                                                },
                                                "end_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The end time must be a date after start time."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка при создании бронирования."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "An internal server error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking by ID",
                "description": "Endpoint to delete an existing booking by its ID.",
                "operationId": "cab4f65d2acdf6c9d4663c896e25e6fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Бронирование успешно удалено."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Бронирование не найдено."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка при удалении бронирования."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message from exception"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}